+++ src/cmd/grit/cd.go
@@ -1,11 +1,8 @@
 package main
 
 import (
-	"fmt"
-
 	"github.com/jmalloc/grit/src/grit"
 	"github.com/jmalloc/grit/src/grit/index"
-	"github.com/jmalloc/grit/src/grit/pathutil"
 	"github.com/urfave/cli"
 )
 
@@ -16,22 +13,13 @@ func cd(cfg grit.Config, idx *index.Index, c *cli.Context) error {
 	}
 
 	dirs := idx.Find(slug)
-	gosrc, _ := pathutil.GoSrc()
-	var opts []string
-
-	for _, dir := range dirs {
-		if rel, ok := pathutil.RelChild(gosrc, dir); ok && gosrc != "" {
-			opts = append(opts, fmt.Sprintf("[go] %s", rel))
-		} else if rel, ok := pathutil.RelChild(cfg.Clone.Root, dir); ok {
-			opts = append(opts, fmt.Sprintf("[grit] %s", rel))
-		} else {
-			opts = append(opts, dir)
-		}
+	if len(dirs) == 0 {
+		return notIndexed(slug)
 	}
 
-	if i, ok := choose(c, opts); ok {
-		write(c, dirs[i])
-		exec(c, "cd", dirs[i])
+	if dir, ok := chooseCloneDir(cfg, c, dirs); ok {
+		write(c, dir)
+		exec(c, "cd", dir)
 		return nil
 	}